Only copy AR17XX.hints if it exists

+++ builder_scripts/builder_common.sh
@@ -260,7 +260,9 @@ fixup_kernel_options() {
 	cp $SRCDIR/sys/$ARCH/conf/GENERIC.hints $KERNEL_BUILD_PATH/uniprocessor/boot/device.hints
 	cp $SRCDIR/sys/$ARCH/conf/GENERIC.hints $KERNEL_BUILD_PATH/SMP/boot/device.hints
 	cp $SRCDIR/sys/$ARCH/conf/GENERIC.hints $KERNEL_BUILD_PATH/developers/boot/device.hints
-	cp $SRCDIR/sys/mips/conf/AR71XX.hints $KERNEL_BUILD_PATH/AR71XX/boot/device.hints
+	if [ -f $SRCDIR/sys/mips/conf/AR71XX.hints ]; then
+		cp $SRCDIR/sys/mips/conf/AR71XX.hints $KERNEL_BUILD_PATH/AR71XX/boot/device.hints
+	fi
 	# END NOTE.
 
 	# Danger will robinson -- 7.2+ will NOT boot if these files are not present.
